Super refactoring commit. Introduced Parsable.
Parsable abstracted nearly everything about the parsing away from the
actual class definitions for the parsable things that I'm concerned
with. It also facilitated the independence of all of the nonterminals
to get semantic content (classes) of their own, and I think it's really
nice now. But it could be nicer. Refactoring this synchronous production
stuff is the last step to making it really beautiful.