Hugely refactored testing to be awesome now.
Introduced the notion of ParsableTestParameters and a common testing
framework for all of the Parsables. Implemented this (some more
bare-bones than others) for the basic Parsables. The next step is to
do it for the schemas and also provide a SchemaTestParameters trait and
appropriate testing mechanism.

Also, I specified grammars for the schema classes. Not sure if I did
that before the last commit. They're Parsable now, it just remains to
test them...