Totally reorganized, and added deduction proofs!
Now there's an individual package `parsing` for the
exclusively-parsing-related stuff, a package `propositional` for the
propositional logic stuff, a subpackage `propositional.schema` for all
of the associated schemas, and a package `util` for extraneous utility
stuff (which as of now only has Memoize). I also added a class for
deduction proofs that can validate a proof for you given a list of
axioms and deduction rules! yay! Also in TestMain is a routine that
could probably be generalized to help with SHOWING the deduction proofs
interactively. Right now it's... not it's own routine but the algorithm
is there to get the string of sequents into a DeductionProof, validate
it, and display the results in a nice readable way. Huzzah, milestone!

Note: Still need testing for the schemas. There was a mistake in my
UnionSchema implementation that I just fixed. But maaaaan, that is not
the most exciting part of this project... :P