Massive changes to testing infrastructure
Incorporated new testing code from parsing, and changed the way we interface
with it. Now each type has its own test suite (as it should be). Also, changed
to scala 2.11.1.

It turns out I was mistaken when I wrote that lexical categories worked. Well, I
was mistaken to say it: It might turn out that they DO work but my test code is
wrong. I still need to look deeply at the tests that are failing to see what's
going wrong. Anyway, the tests FINALLY compile and run and mostly succeed (at
least, where expected) so we're making progress.

Once lexical categories are confirmed to work, I'm planning on working more on
the parsing package alone. I'm going to do that by writing test code in that
package to test its own functionality instead of only doing these E2E type tests
on this end. In fact, I might do that before taking a closer look at these
tests, because it would make sense to first test the parsing code with simpler
examples. Maybe there'd be some useful parsable stuff to put in the package that
can serve that purpose: for example, my own int and double parsers, or that sort
of thing.