TESTS ALL PASSING! Introduced a hack so Word works
Finally got everything in order so all of the tests pass and everything works in
the paradigm of LexicalCategory. However, we might still have a problem. The
issue is that the object Word is supposed to be a parsable that is not any of
the other terminals. This means that information about the other terminals must
be somehow referenced by its subLexicon function. There's no facility for this
and I don't see the facility to abstract this out: parsables like terminals and
lexical categories should not be aware of each other unless they have a more
complicated relationship. Maybe this shows that I need a different model
altogether, and all of the lexical categories should be grouped together and
allowed to reference each other and build off of each other. Indeed, this is
probably what I'll end up with if I go in the direction of natural language and
have transformations from some parts of speech to others. I will delay the
improvement of this abstraction until then, as it is serving well enough for the
time being.